Refactor: rename useInitializeChat to useChat and move all sendbird logic to that hook
